# docs/generated/backend.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Транспорт · Backend",
    "description": "Диспетчерская система: demo, отдельный synthetic generator, live NDTP и CSV replay. Отдельный ML-сервис с frozen-моделью; при недоступности — явный persistence fallback. Времена RFC3339 с offset.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/demo/control": {
      "post": {
        "tags": [
          "Демонстрация"
        ],
        "summary": "Control",
        "operationId": "control_api_v1_demo_control_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoControl"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mode": {
      "post": {
        "tags": [
          "Демонстрация"
        ],
        "summary": "Mode",
        "description": "Переключение очищает старые координаты, расписание, прогнозы и алерты.",
        "operationId": "mode_api_v1_mode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeControl"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replay/load": {
      "post": {
        "tags": [
          "Исторический replay"
        ],
        "summary": "Replay Load",
        "description": "Загрузить validate или train из REPLAY_DATA_DIR; train только в режиме gps.\n\nИз train/schedule.csv используются только плановые поля. labels и\nфактические прибытия не становятся входами детектора или модели.\n\ntr_ids=null выбирает все ТС с планом и сообщениями в срезе с прогревом.\nЯвный список — ручной выбор до 128 ТС. Счётчики и будущие времена\nв replay.streams служат только диагностикой очереди проигрывателя.\n\nduration_minutes=null (по умолчанию) читает до последней доставки\nвыбранного архива: max(event_time, receive_time) для GPS, T для CSV.\nЯвное число ограничивает окно сверху. Пустой хвост не проигрывается.\nПо умолчанию начинает с паузы. Загрузка очищает текущий контекст только\nпосле успешной проверки CSV. Пути файлов через API не задаются.",
        "operationId": "replay_load_api_v1_replay_load_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/generator/start": {
      "post": {
        "tags": [
          "Сценарный генератор"
        ],
        "summary": "Generator Start",
        "description": "Новая сессия отдельного HTTP producer; план синтетический, будущих фактов нет.\n\nПараметры воспроизводимы через seed. Сначала проверяется ответ producer,\nзатем заменяется контекст. URL задаётся только конфигурацией сервера.",
        "operationId": "generator_start_api_v1_generator_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratorStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/generator/control": {
      "post": {
        "tags": [
          "Сценарный генератор"
        ],
        "summary": "Generator Control",
        "description": "Пауза/скорость меняют часы producer. Backend забирает все доступные кадры по cursor.",
        "operationId": "generator_control_api_v1_generator_control_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratorControl"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/replay/control": {
      "post": {
        "tags": [
          "Исторический replay"
        ],
        "summary": "Replay Control",
        "operationId": "replay_control_api_v1_replay_control_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayControl"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/telemetry": {
      "post": {
        "tags": [
          "Входные данные"
        ],
        "summary": "Telemetry",
        "description": "HTTP-адаптер decoded CSV/других источников; при приёме received_at ставит сервер.",
        "operationId": "telemetry_api_v1_telemetry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/Telemetry"
                },
                "type": "array",
                "maxItems": 500,
                "minItems": 1,
                "title": "Points"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/arrivals": {
      "post": {
        "tags": [
          "Входные данные"
        ],
        "summary": "Arrival",
        "description": "Факт остановки от отдельного источника. NDTP Nav00 такого события не даёт.\n\nBackend вычисляет cur_dev_s = arrived_at − scheduled_at. Повтор того же\nфакта идемпотентен, конфликт времени — 409; будущий факт/неизвестная ссылка — 422.\nОперационные факты поступают сюда отдельно. Экспериментальный внутренний\nGPS-детектор включается через arrival_mode=gps; подтверждённый операционный\nфакт может уточнить его оценку, сохраняя время доступности обеих версий.",
        "operationId": "arrival_api_v1_arrivals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArrivalInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles/{tr_id}/forecast-trace": {
      "get": {
        "tags": [
          "Диспетчер"
        ],
        "summary": "Forecast Trace",
        "description": "Последний завершённый запрос/ответ инференса с происхождением current_delay.\n\nРезультат может быть историческим: issued_at и published_at нужно проверять.\nПри новом контексте очищается. Ответ не содержит будущих фактов прибытия.",
        "operationId": "forecast_trace_api_v1_vehicles__tr_id__forecast_trace_get",
        "parameters": [
          {
            "name": "tr_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tr Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastTrace"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DemoControl": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "pause",
              "resume",
              "reset",
              "speed",
              "source_off",
              "source_on"
            ],
            "title": "Action"
          },
          "speed": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 30.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action"
        ],
        "title": "DemoControl"
      },
...
      "GeneratorControl": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "pause",
              "resume",
              "speed"
            ],
            "title": "Action"
          },
          "speed": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 30.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action"
        ],
        "title": "GeneratorControl"
      },
      "GeneratorStart": {
        "properties": {
          "scenario": {
            "type": "string",
            "enum": [
              "normal",
              "slow_segment",
              "long_stop",
              "gps_loss"
            ],
            "title": "Scenario",
            "default": "normal"
          },
          "seed": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Seed",
            "default": 42
          },
          "speed": {
            "type": "number",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Speed",
            "default": 10
          },
          "paused": {
            "type": "boolean",
            "title": "Paused",
            "default": true
          },
          "route_count": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Route Count",
            "default": 4
          },
          "telemetry_interval_s": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Telemetry Interval S",
            "default": 15
          },
          "door_sensors": {
            "type": "boolean",
            "title": "Door Sensors",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "GeneratorStart"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModeControl": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "live"
            ],
            "title": "Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "ModeControl"
      },
...
      "ReplayConfig": {
        "properties": {
          "dataset_split": {
            "type": "string",
            "enum": [
              "validate",
              "train"
            ],
            "title": "Dataset Split",
            "default": "validate"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start",
            "default": "2026-01-06T11:30:00Z"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes",
            "description": "null: до последнего доступного события архива; число: верхняя граница окна для короткой проверки"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "UTC"
          },
          "tr_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array",
                "maxItems": 128,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Tr Ids",
            "description": "null: все ТС с планом и сообщениями в срезе с предысторией; список: ручной выбор"
          },
          "speed": {
            "type": "number",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Speed",
            "default": 10
          },
          "paused": {
            "type": "boolean",
            "title": "Paused",
            "default": true
          },
          "deviation_source": {
            "type": "string",
            "enum": [
              "csv_snapshot",
              "gps"
            ],
            "title": "Deviation Source",
            "default": "csv_snapshot"
          },
          "warmup_minutes": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 0.0,
            "title": "Warmup Minutes",
            "description": "Сколько доступной истории перед start доставить для прогрева; будущие события остаются в очереди",
            "default": 5
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ReplayConfig"
      },
...
    }
  }
}